requirements.txt
pillow
torch
clip
transformers
open_clip_torch
gradio
numpy
nltk